+++ apps/admin-x-settings/src/components/providers/routing/modals.tsx
@@ -4,12 +4,12 @@ import type {RoutingModalProps} from '../RoutingProvider';
 import AboutModal from '../../settings/general/About';
 import AddIntegrationModal from '../../settings/advanced/integrations/AddIntegrationModal';
 import AddNewsletterModal from '../../settings/email/newsletters/AddNewsletterModal';
-import AddRecommendationModal from '../../settings/site/recommendations/AddRecommendationModal';
+import AddRecommendationModal from '../../settings/membership/recommendations/AddRecommendationModal';
 import AmpModal from '../../settings/advanced/integrations/AmpModal';
 import AnnouncementBarModal from '../../settings/site/AnnouncementBarModal';
 import CustomIntegrationModal from '../../settings/advanced/integrations/CustomIntegrationModal';
 import DesignAndThemeModal from '../../settings/site/DesignAndThemeModal';
-import EditRecommendationModal from '../../settings/site/recommendations/EditRecommendationModal';
+import EditRecommendationModal from '../../settings/membership/recommendations/EditRecommendationModal';
 import EmbedSignupFormModal from '../../settings/membership/embedSignup/EmbedSignupFormModal';
 import FirstpromoterModal from '../../settings/advanced/integrations/FirstPromoterModal';
 import HistoryModal from '../../settings/advanced/HistoryModal';
